refactor: split upgrade repo script and add tests
